refactor: move things into src
