+++ jest.config.js
@@ -2,8 +2,8 @@ module.exports = {
   testEnvironment: 'node',
   coverageDirectory: 'coverage',
   collectCoverageFrom: [
-    'scripts/lib/**/*.js',
-    '!scripts/lib/**/*.test.js'
+    'src/**/*.js',
+    '!src/**/*.test.js'
   ],
   testMatch: [
     '**/tests/**/*.test.js'
