+++ src/error-recovery-manager.js
