+++ src/git-lfs-detector.js
