+++ src/image-optimizer.js
