+++ src/image-processor.js
