+++ src/output-path-generator.js
