+++ src/progress-manager.js
