+++ tests/lib/config-integration.test.js
@@ -1,7 +1,7 @@
 const fs = require('fs').promises;
 const path = require('path');
 const os = require('os');
-const ConfigLoader = require('../../scripts/lib/config-loader');
+const ConfigLoader = require('../../src/config-loader');
 
 describe('Configuration Integration', () => {
   let tempDir;
